ccl(provenance): declare the pipeline's pane topology in one place
Three panes and two pane pairs were spelled across six consts, six named
`MaterializedPanes` fields and two methods, which does not survive going to six
panes: five pairs times three fold products is fifteen fields. `PANES` now
declares the topology once — each pane's name, the phases that produced it from
the pane before it, and whether its pair is gated — and `materialize_panes`
folds a sliding window over it. `INFER_PHASES` and `CHANNELIZE_PHASES` fold into
that table; `MaterializedPanes` becomes a projection per pane and a `PanePair`
per adjacent pair, each pair naming itself by joining its two pane names so a
name cannot drift from the topology.

All of it moves to `src/ccl/panes.rs`, along with `gate_leaks`, the two
`CompiledProgram` methods that fold and enumerate the panes, and the pane tests.
`context.rs` loses a third of its length and keeps what it is about: the
pipeline, the `Phase` axis, and the env switches. The inherent `impl
CompiledProgram` sits in `panes.rs` rather than beside the struct because the
methods are the pane layer's, which is also why no call site had to change.

The doc's "The seam" heading drops its `(src/ccl/context.rs)` suffix, since two
of the section's four bullets now describe `panes.rs`; the three existing
citations of it follow. A path in a heading is a thing to keep current, and this
is the second time it would have needed it.

Two panes are added, both below the ones that existed. `post-as-of-read` is
taken after `transact_phase::rewrite_as_of_reads` and `post-lambda-elim` after
`lambda_elim::run`; the existing `ast` is registered as `post-planning` rather
than retained twice.

`Phase::AsOfRead` is new, declared between `Channelize` and `LambdaElim` because
that is where the rewrite runs. Its code lives in `transact_phase`, but the
module a rewrite's code sits in is not its phase: reusing `Phase::Transact`
would have put one phase on both sides of the post-channelize pane, and a fold
restricting by it would resolve straight past that pane. The scope is opened
here, so the rewrite's existing `transact.as_of_read` recording reaches a table
in a normal compile for the first time — `post-channelize → post-as-of-read` is
zero on both leak classes over the corpus, and shows up as an exercised pair on
the transaction fixture.

`PaneSpec::gated` stays one bit per pair, and it says whether an assertion is
turned on rather than how clean a pair is allowed to be — a leak is a bug
wherever it appears. Two checks read the flag at different widths:
`pane_pair_folds_have_no_structural_leaks` over `corpus()` on every test run, and
`CAMBRA_PROVENANCE_GATE` over whatever the caller compiles, which is strictly
stronger because the wider corpus reaches shapes the eleven listed programs do
not. `pre-inference → post-inference` is gated for the first time, the two
commits below having closed both gaps that held it back; `post-inference →
post-channelize` stays gated; `post-channelize → post-as-of-read` is gated, the
as-of-read rewrite's existing recording being total once a scope exists to land
it in. The two pairs below stay ungated until `lambda_elim` and `planning`
record: what an uninstrumented phase leaves behind is a count of how little it
records, a constant no correct recording elsewhere can drive down, so gating it
would pin churn rather than catch a defect.

There is no `PANE_COUNT`. `PANES.len()` works as an array length, so
`pane_trees`' arity is `PANES`' own rather than a third constant both have to
match. `MaterializedPanes::pane_pairs()` is not added either — `pairs` is
`pub(crate)` and the accessor would have no caller.

The corpus test's projection check now runs over the gated prefix rather than
every pane: a pair folds against the projection of the pane above it, so the
first pair whose phases do not record leaves every pane below it with nothing to
carry forward, and an empty projection there is the expected state.

`every_recorded_phase_belongs_to_exactly_one_pane_pair` generalizes from "in one
of two sets" to a count over `PANES`, which is what it always meant.
