+++ src/ccl/context.rs
@@ -18,10 +18,12 @@ use crate::{
         },
         inline, lambda_elim,
         lower::{LoweringContext, LoweringError, lower_stmts},
-        mut_elim, planning,
+        mut_elim,
+        panes::gate_leaks,
+        planning,
         provenance::{
-            Leak, LoweringSession, NodeId, PhaseScope, ProvenanceMap, ProvenanceTable,
-            SourceProjection, TableSession, fold, fold_lowering,
+            Leak, LoweringSession, NodeId, PhaseScope, ProvenanceTable, SourceProjection,
+            TableSession, fold, fold_lowering,
         },
         symbolic::{symbolic, symbolic_typed},
         transact_phase, uniquify,
@@ -534,6 +536,22 @@ pub struct CompiledProgram {
     /// Every id preserved through inline/transact/letrec/channelize is shared with
     /// [`post_inference_ir`](Self::post_inference_ir).
     pub post_channelize_ir: Expr,
+    /// The post-as-of-read IR snapshot, taken after
+    /// [`transact_phase::rewrite_as_of_reads`](crate::ccl::transact_phase::rewrite_as_of_reads).
+    ///
+    /// Distinct from [`post_channelize_ir`](Self::post_channelize_ir) because
+    /// that rewrite runs between the two: on a program with a fed-out mutable
+    /// read the trees differ, and on every other program they agree node for
+    /// node. Isolating it is what lets the pair ending here gate that one
+    /// rewrite rather than fold it in with lambda elimination.
+    pub post_as_of_read_ir: Expr,
+    /// The post-lambda-elim IR snapshot, taken after
+    /// [`lambda_elim::run`](crate::ccl::lambda_elim::run).
+    ///
+    /// The pass re-mints nearly every pass-through node, so almost no id here is
+    /// shared with the pane before it and the pair's relation falls to the
+    /// parents walk rather than to shared ids. That is what the pair measures.
+    pub post_lambda_elim_ir: Expr,
     /// The compile's provenance record: `NodeId → { parents, blame, rule }`, one
     /// row per node a phase produced, written by the recorder as those phases run
     /// ([`crate::ccl::provenance`]).
@@ -544,8 +562,8 @@ pub struct CompiledProgram {
     /// [`Phase::Infer`] (everything monomorphization mints inside `infer`, which
     /// bridges the pre-inference ⇄ post-inference panes) and [`Phase::Inline`],
     /// [`Phase::Transact`], [`Phase::Letrec`], [`Phase::Channelize`] (the four phases
-    /// between the post-inference and post-channelize snapshots) — see
-    /// [`INFER_PHASES`] and [`CHANNELIZE_PHASES`].
+    /// between the post-inference and post-channelize snapshots) — see [`PANES`],
+    /// which declares which phases each pane pair folds.
     ///
     /// Rows exist only for nodes a recording produced, so an untouched node has
     /// none — it was never rewritten. Refinement-predicate interiors are rows
@@ -606,57 +624,6 @@ impl CompiledProgram {
     pub fn sinks(&self) -> impl Iterator<Item = &CompiledOutput> {
         self.outputs.iter().filter(|o| !o.is_main())
     }
-
-    /// Fold [`provenance_table`](Self::provenance_table) across the two pane pairs into
-    /// the per-pane [`SourceProjection`]s, the pane-pair [`ProvenanceMap`]s, and
-    /// each pair's [`Leak`]s. Cold path (snapshot-serve only), never called
-    /// by [`compile_program`]:
-    ///
-    /// * pre-inference pane = the lowering projection (`uniquify` preserves every
-    ///   id in place, so lowering's keys are still the pane's keys);
-    /// * post-inference pane = fold the [`INFER_PHASES`] rows against the
-    ///   pre-inference pane;
-    /// * post-channelize pane = fold the [`CHANNELIZE_PHASES`] rows against the
-    ///   post-inference pane.
-    ///
-    /// The leaks are **returned, not asserted**: a span reaches zero only once
-    /// every phase inside it records its rewrites, so [`gate_leaks`] is left to
-    /// the callers that fold an instrumented span. The deaths ride alongside them
-    /// as a product, since nothing declares a fate.
-    // Cold path: the inspector's snapshot serve, which is not in this workspace.
-    #[allow(dead_code)]
-    pub(crate) fn materialize_panes(&self) -> MaterializedPanes {
-        let pre_ids = collect_tree_ids(&self.pre_inference_ir);
-        let post_inf_ids = collect_tree_ids(&self.post_inference_ir);
-        let post_des_ids = collect_tree_ids(&self.post_channelize_ir);
-
-        let (infer_map, post_inference, infer_deaths, infer_leaks) = fold(
-            &self.provenance_table,
-            INFER_PHASES,
-            &pre_ids,
-            &post_inf_ids,
-            &self.lowering_projection,
-        );
-        let (channelize_map, post_channelize, channelize_deaths, channelize_leaks) = fold(
-            &self.provenance_table,
-            CHANNELIZE_PHASES,
-            &post_inf_ids,
-            &post_des_ids,
-            &post_inference,
-        );
-
-        MaterializedPanes {
-            pre_inference: self.lowering_projection.clone(),
-            post_inference,
-            post_channelize,
-            infer_map,
-            channelize_map,
-            infer_deaths,
-            channelize_deaths,
-            infer_leaks,
-            channelize_leaks,
-        }
-    }
 }
 
 /// The compiler phase that rewrote a node — a
@@ -716,102 +683,23 @@ pub enum Phase {
     /// synthesizes new nodes (channel unions, contribution records, floated
     /// lambdas, DI wrappers) that are tagged `{via: Channelize, nature: Machinery}`.
     Channelize,
+    /// The as-of-read rewrite
+    /// ([`crate::ccl::transact_phase::rewrite_as_of_reads`]): turning a
+    /// read-only reply that reads a mutable variable out of its block into an
+    /// outer-indexed as-of join.
+    ///
+    /// Its code lives in `transact_phase`, but a phase is declared where the
+    /// rewrite *runs*, not where its code sits — this one runs after
+    /// `Channelize` and before `LambdaElim`. Reusing [`Transact`](Self::Transact)
+    /// for it would put one phase on both sides of the post-channelize pane, and
+    /// a fold restricting by that phase would resolve straight past it.
+    AsOfRead,
     /// Lambda elimination: synthesizing point-free combinators (`Compose`,
-    /// `Zip`, `Id`) from explicit lambdas. **Never constructed**: `lambda_elim`
-    /// opens no recording, so no row carries this tag. The variant exists so the
-    /// axis covers every phase.
+    /// `Zip`, `Id`) from explicit lambdas.
     LambdaElim,
     /// Join/dataflow planning: hash-join and restrict scaffolding, clause
     /// fusion, refinement-predicate compilation.
     Planning,
-}
-
-/// The phases between each adjacent pair of panes — the set
-/// [`CompiledProgram::materialize_panes`] restricts the whole-compile table by.
-///
-/// A pane pair's fold is defined by the phases that ran between its two panes,
-/// not by a position in a list: a program that skips a phase must not shift the
-/// other pair's set, and a row produced outside a pair's phases has to read to it
-/// as an ordinary un-produced id.
-pub(crate) const INFER_PHASES: &[Phase] = &[Phase::Infer];
-
-/// The phases between the post-inference and post-channelize panes. See
-/// [`INFER_PHASES`].
-pub(crate) const CHANNELIZE_PHASES: &[Phase] = &[
-    Phase::Inline,
-    Phase::Transact,
-    Phase::Letrec,
-    Phase::Channelize,
-];
-
-/// The per-pane projections, pane-pair provenance maps, and per-pair leaks
-/// materialized from [`CompiledProgram::provenance_table`] — see
-/// [`CompiledProgram::materialize_panes`].
-// Consumed by the inspector model; unused within the compiler itself.
-#[allow(dead_code)]
-pub(crate) struct MaterializedPanes {
-    /// pre-inference pane projection (= the lowering projection).
-    pub(crate) pre_inference: SourceProjection,
-    /// post-inference pane projection.
-    pub(crate) post_inference: SourceProjection,
-    /// post-channelize pane projection.
-    pub(crate) post_channelize: SourceProjection,
-    /// pre-inference → post-inference provenance map: `Infer`'s rewrites, of
-    /// which monomorphization's fan-out is the bulk. Dense: an id that survived is
-    /// its own self-edge.
-    pub(crate) infer_map: ProvenanceMap<NodeId, NodeId>,
-    /// post-inference → post-channelize provenance map.
-    pub(crate) channelize_map: ProvenanceMap<NodeId, NodeId>,
-    /// The pre-inference ids absent from the post-inference pane — `input_ids ∖
-    /// output_ids`, which is the whole of what "died" means here.
-    pub(crate) infer_deaths: Vec<NodeId>,
-    /// The post-inference ids absent from the post-channelize pane.
-    pub(crate) channelize_deaths: Vec<NodeId>,
-    /// Leaks between the pre-inference and post-inference panes.
-    pub(crate) infer_leaks: Vec<Leak>,
-    /// Leaks between the post-inference and post-channelize panes.
-    pub(crate) channelize_leaks: Vec<Leak>,
-}
-
-impl MaterializedPanes {
-    /// `(pane-pair name, leaks)` for each pane pair, in pipeline order.
-    #[allow(dead_code)]
-    pub(crate) fn pane_pairs(&self) -> [(&'static str, &[Leak]); 2] {
-        [
-            ("pre-inference → post-inference", &self.infer_leaks),
-            ("post-inference → post-channelize", &self.channelize_leaks),
-        ]
-    }
-
-    /// The pane pairs a gate can hold at zero: the ones whose phases are
-    /// **instrumented**. Same discipline as an audit span's endpoint. A gate over
-    /// a pane pair whose phases do not record cannot reach zero however correct
-    /// the recording is, so it would report a constant rather than a regression.
-    ///
-    /// Today that is the second pair only. The first spans monomorphization and
-    /// inference, and **inference's predicate producers do not record**:
-    /// `specialize_use` clones a definition per instantiation, and the copies of
-    /// a predicate term inside it row against interior ids no phase ever
-    /// produced. It is the crossing "Known prerequisites" records.
-    ///
-    /// The residue's class signature is what makes it a constant. Inference's own
-    /// rewrites sit inside a recording scope, so every post-inference node
-    /// carries a row and [`Leak::Unrecorded`] is zero; what dangles is the
-    /// parents those rows name. Returning both pairs here and running
-    /// `CAMBRA_PROVENANCE_GATE=1` over `tests/compilation_pipeline` fails six
-    /// programs on 46 leaks, every one [`Leak::DanglingParent`] and none
-    /// [`Leak::Unrecorded`]: four UDF-with-filter or poly-wrapper shapes and two
-    /// refined-annotation shapes. A [`Leak::Unrecorded`] arriving there would be
-    /// a missed mint and a defect now, which is the reading the two classes carry
-    /// that a single class would not.
-    ///
-    /// No count is pinned here, because a count over an uninstrumented phase
-    /// measures how little it records and churns with every unrelated change.
-    /// **Add the first pair here in the commit that makes inference record**,
-    /// exactly as an audit span's endpoint moves with the phase that earns it.
-    pub(crate) fn gated_pane_pairs(&self) -> [(&'static str, &[Leak]); 1] {
-        [("post-inference → post-channelize", &self.channelize_leaks)]
-    }
 }
 
 /// Ids carried by two *distinct* predicate terms, or by a predicate term and the
@@ -955,26 +843,6 @@ pub(crate) fn collect_tree_ids(expr: &Expr) -> std::collections::HashSet<NodeId>
     acc
 }
 
-/// The leak gate: **a fold's [`Leak`] vector must be empty**.
-///
-/// Both classes are asserted the same way. Each means a node reached the output
-/// pane with nothing recording where it came from, and neither localizes the site
-/// to fix on its own, so the gate reads the vector's emptiness and not its
-/// composition (see [`Leak`]). A death is not a leak and reaches a caller as its
-/// own collection, so there is nothing here to filter.
-///
-/// Debug/test only, single code path (`cfg!`, not `#[cfg]`).
-fn gate_leaks(leaks: &[Leak], pair: &str) {
-    if !cfg!(any(debug_assertions, test)) {
-        return;
-    }
-    assert!(
-        leaks.is_empty(),
-        "provenance capture defect between the {pair} panes ({} leaks): {leaks:?}",
-        leaks.len(),
-    );
-}
-
 /// Every duplicated [`NodeId`] over the **main tree** — the `walk_children`
 /// node-set, refinement predicates excluded. Returns `(id, node kind)` for each
 /// occurrence *beyond the first*.
@@ -1091,7 +959,7 @@ const PROVENANCE_PREDICATES_ENV: &str = "CAMBRA_PROVENANCE_PREDICATES";
 /// `#[test]`, so the name is dead in a lib build; it lives here rather than
 /// beside its reader so that adding a fifth switch means editing one block.
 #[cfg(test)]
-const PERF_REPS_ENV: &str = "CAMBRA_PERF_REPS";
+pub(crate) const PERF_REPS_ENV: &str = "CAMBRA_PERF_REPS";
 
 /// The audit span named by [`PROVENANCE_AUDIT_ENV`], if any. **Sole reader** of
 /// that variable — see the section note above.
@@ -1642,15 +1510,21 @@ pub fn compile_program(
     // rather than a point-free `const` a planning-time recognizer would have to
     // reject. Uniform across the reading loop's domain. See
     // `transact_phase::rewrite_as_of_reads`.
-    transact_phase::rewrite_as_of_reads(&mut channelized)
-        .map_err(|msg| vec![CompileError::Unsupported(msg)])?;
+    recorded(capture_provenance, Phase::AsOfRead, || {
+        transact_phase::rewrite_as_of_reads(&mut channelized)
+    })
+    .map_err(|msg| vec![CompileError::Unsupported(msg)])?;
     assert_unique_node_ids(&channelized, "post-as-of-read");
     typecheck(&channelized).expect("as-of-read rewrite produced an ill-typed tree");
     // The last instrumented pane: see the span's own note at `ProvenanceAudit::start`.
     audit.finish(&channelized);
+    // A pane snapshot; see `pre_inference_ir`.
+    let post_as_of_read_ir = channelized.clone_preserving_ids();
 
     let lambda_elim = lambda_elim::run(channelized).errs()?;
     assert_unique_node_ids(&lambda_elim, "post-lambda-elim");
+    // A pane snapshot; see `pre_inference_ir`.
+    let post_lambda_elim_ir = lambda_elim.clone_preserving_ids();
     debug!("λ-eliminated CCL:\n{}", symbolic(&lambda_elim));
     debug!("λ-eliminated typed CCL:\n{}", symbolic_typed(&lambda_elim));
 
@@ -1787,6 +1661,8 @@ pub fn compile_program(
         pre_inference_ir,
         post_inference_ir,
         post_channelize_ir,
+        post_as_of_read_ir,
+        post_lambda_elim_ir,
         provenance_table,
         source_ast: module,
         source: code.to_string(),
@@ -1795,8 +1671,8 @@ pub fn compile_program(
     // Every compile its own gate — see `provenance_gate_every_compile` for why this
     // is opt-in and what it covers that the sampled gate does not.
     if provenance_gate_every_compile() {
-        for (pair, leaks) in program.materialize_panes().gated_pane_pairs() {
-            gate_leaks(leaks, pair);
+        for pair in program.materialize_panes().gated_pane_pairs() {
+            gate_leaks(&pair.leaks, &pair.name);
         }
     }
 
@@ -1807,641 +1683,6 @@ pub fn compile_program(
 mod tests {
     use super::*;
     use rstest::rstest;
-
-    /// The pane-measurement corpus: every demo-gallery program that compiles
-    /// today, plus four inline programs covering the phases the gallery does not
-    /// reach (a `with begin():` transaction, a group-by, a UDF chain, and a
-    /// nested comprehension).
-    ///
-    /// The gallery's remaining programs are excluded for reasons unrelated to
-    /// provenance: most are deliberate *failure* fixtures (`while`, record-term
-    /// syntax, `Feed(_)` types) that pin errors and so have no panes to fold,
-    /// and the three HTTP demos bind a real listening socket during lowering,
-    /// which collides with itself under a parallel test runner.
-    fn corpus() -> Vec<(&'static str, String)> {
-        vec![
-            (
-                "arithmetic",
-                include_str!("../../tests/programs/arithmetic/program.cambra").to_string(),
-            ),
-            (
-                "filter_and_aggregate",
-                include_str!("../../tests/programs/filter_and_aggregate/program.cambra").to_string(),
-            ),
-            (
-                "for_accumulator",
-                include_str!("../../tests/programs/for_accumulator/program.cambra").to_string(),
-            ),
-            (
-                "generator_pipeline",
-                include_str!("../../tests/programs/generator_pipeline/program.cambra").to_string(),
-            ),
-            (
-                "inner_join",
-                include_str!("../../tests/programs/inner_join/program.cambra").to_string(),
-            ),
-            (
-                "prefix_lines",
-                include_str!("../../tests/programs/prefix_lines/program.cambra").to_string(),
-            ),
-            (
-                "streaming_echo",
-                include_str!("../../tests/programs/streaming_echo/program.cambra").to_string(),
-            ),
-            (
-                "transaction",
-                "out = defer()\n\
-                 pool: Mut(Int, Txn) := 100\n\
-                 for r in [10, 20, 30]:\n\
-                 \x20   with begin():\n\
-                 \x20       pool := pool - r\n\
-                 with begin():\n\
-                 \x20   out << pool\n\
-                 out\n"
-                    .to_string(),
-            ),
-            (
-                "group_by",
-                "[sum(x) for x in groupby([y + 10 for y in [2,3,4,5,6] if y < 6], \\x -> x // 2)]\n"
-                    .to_string(),
-            ),
-            (
-                "udf_chain",
-                "def double(x):\n    x * 2\ndef bump(x):\n    double(x) + 1\n\
-                 xs = [1, 2, 3]\n[bump(x) for x in xs]\n"
-                    .to_string(),
-            ),
-            (
-                "feed_loop",
-                "out = defer()\nfor x in [1, 2, 3]:\n    out << x * 2\nout\n".to_string(),
-            ),
-        ]
-    }
-
-    /// Compile `code` through the full pipeline, panicking on error.
-    fn compile_ok(code: &str) -> CompiledProgram {
-        let mut ctx = GlobalContext::default();
-        let consumer: Box<dyn Consumer> = Box::new(|| {});
-        match compile_program(&mut ctx, code, consumer) {
-            Ok(p) => p,
-            Err(errs) => panic!("expected a successful compile, got {errs:?}"),
-        }
-    }
-
-    /// Per-leak-class counts, for reporting one fold.
-    #[derive(Default, Debug, PartialEq, Eq)]
-    struct LeakCounts {
-        unrecorded: usize,
-        dangling_parent: usize,
-    }
-
-    impl LeakCounts {
-        fn tally(leaks: &[Leak]) -> Self {
-            let mut c = LeakCounts::default();
-            for l in leaks {
-                match l {
-                    Leak::Unrecorded { .. } => c.unrecorded += 1,
-                    Leak::DanglingParent { .. } => c.dangling_parent += 1,
-                }
-            }
-            c
-        }
-
-        fn add(&mut self, o: &LeakCounts) {
-            self.unrecorded += o.unrecorded;
-            self.dangling_parent += o.dangling_parent;
-        }
-    }
-
-    /// **Capture totality**, as a corpus-wide property rather than a per-phase
-    /// assertion: both pane pairs materialize and fold over every corpus
-    /// program, every output-pane node has an origin (`Unrecorded == 0`), and
-    /// no node's ancestry dangles (`DanglingParent == 0`).
-    ///
-    /// The two invariants fail differently and are worth reading apart.
-    /// `dangling_parent == 0` says no node's ancestry stops at an id the fold
-    /// never heard of. `Unrecorded == 0` says no rewrite went
-    /// *unrecorded* — it is the gate the whole driver-capture design exists to
-    /// phase, and the number a newly-added rewrite site breaks first.
-    #[test]
-    fn pane_pair_folds_have_no_structural_leaks() {
-        let mut totals = [LeakCounts::default(), LeakCounts::default()];
-        for (name, code) in corpus() {
-            let program = compile_ok(&code);
-            let panes = program.materialize_panes();
-            for (i, (pair, leaks)) in panes.pane_pairs().into_iter().enumerate() {
-                let c = LeakCounts::tally(leaks);
-                assert_eq!(
-                    c.dangling_parent, 0,
-                    "{name}: structural leaks between the {pair} panes: {c:?}"
-                );
-                assert_eq!(
-                    c.unrecorded, 0,
-                    "{name}: unrecorded output nodes between the {pair} panes — a rewrite \
-                     that mints with nothing recording: {c:?}"
-                );
-                eprintln!("[pane {name} / {pair}] {c:?}");
-                totals[i].add(&c);
-            }
-            // The panes are the thing being materialized: each projection must
-            // hold entries for the tree it describes, and each map edges.
-            assert!(!panes.pre_inference.is_empty(), "{name}");
-            assert!(!panes.post_inference.is_empty(), "{name}");
-            assert!(!panes.post_channelize.is_empty(), "{name}");
-            assert!(!panes.infer_map.edges().is_empty(), "{name}");
-            assert!(!panes.channelize_map.edges().is_empty(), "{name}");
-        }
-        eprintln!("[pane totals] pre→post-inference {:?}", totals[0]);
-        eprintln!(
-            "[pane totals] post-inference→post-channelize {:?}",
-            totals[1]
-        );
-    }
-
-    /// Every phase that records rows in a normal compile belongs to exactly one
-    /// pane pair, so no rewrite is folded twice and none is silently dropped.
-    ///
-    /// [`INFER_PHASES`] and [`CHANNELIZE_PHASES`] are the only things that decide
-    /// which pane pair a row reaches, so a phase that opens a scope without
-    /// joining neither would record rows nothing ever folds.
-    #[test]
-    fn every_recorded_phase_belongs_to_exactly_one_pane_pair() {
-        for (name, code) in corpus() {
-            let program = compile_ok(&code);
-            for p in program.provenance_table.recorded_phases() {
-                assert!(
-                    INFER_PHASES.contains(&p) != CHANNELIZE_PHASES.contains(&p),
-                    "{name}: {p:?} is in neither pane pair, or in both",
-                );
-            }
-        }
-    }
-
-    /// The `program / pane pair` names at which the provenance map is **not**
-    /// vacuous — where some node's origin is another node, so the fold had to
-    /// read a row. See
-    /// [`the_pane_folds_derive_a_non_vacuous_provenance_map`].
-    ///
-    /// Pinned rather than counted: a recording that opens after the clone that
-    /// produced its nodes captures nothing, and the fold still succeeds — it
-    /// reads the pane pair as pure identity. The pin is what makes that
-    /// difference visible, so it lists names and not a total.
-    const EXERCISED_BOUNDARIES: &[&str] = &[
-        "arithmetic / pre-inference → post-inference",
-        "feed_loop / post-inference → post-channelize",
-        "feed_loop / pre-inference → post-inference",
-        "filter_and_aggregate / pre-inference → post-inference",
-        "for_accumulator / post-inference → post-channelize",
-        "for_accumulator / pre-inference → post-inference",
-        "generator_pipeline / post-inference → post-channelize",
-        "generator_pipeline / pre-inference → post-inference",
-        "group_by / pre-inference → post-inference",
-        "inner_join / pre-inference → post-inference",
-        "prefix_lines / pre-inference → post-inference",
-        "streaming_echo / pre-inference → post-inference",
-        "transaction / post-inference → post-channelize",
-        "transaction / pre-inference → post-inference",
-        "udf_chain / post-inference → post-channelize",
-        "udf_chain / pre-inference → post-inference",
-    ];
-
-    /// **The provenance map is well-formed and non-vacuous** on every corpus
-    /// program: every edge runs from an input-pane id to an output-pane id,
-    /// every id present in both panes is its own dense self-edge, and the
-    /// pane pairs where the fold had to read a *row* are exactly the ones
-    /// pinned above.
-    ///
-    /// The non-vacuity half is what this test is for. Six of the twenty-two
-    /// corpus pane pairs are pure identity: no phase inside them minted, so the
-    /// map is the input pane's self-edges and holds for reasons that have
-    /// nothing to do with the recording. A corpus edit that
-    /// silently dropped the rewriting programs would otherwise leave a green
-    /// tautology behind.
-    #[test]
-    fn the_pane_folds_derive_a_non_vacuous_provenance_map() {
-        let mut exercised: Vec<String> = Vec::new();
-        for (name, code) in corpus() {
-            let program = compile_ok(&code);
-            let panes = program.materialize_panes();
-            let pre = collect_tree_ids(&program.pre_inference_ir);
-            let post_inf = collect_tree_ids(&program.post_inference_ir);
-            let post_des = collect_tree_ids(&program.post_channelize_ir);
-
-            let pairs: [(&str, &ProvenanceMap<NodeId, NodeId>, _, _); 2] = [
-                (
-                    "pre-inference → post-inference",
-                    &panes.infer_map,
-                    &pre,
-                    &post_inf,
-                ),
-                (
-                    "post-inference → post-channelize",
-                    &panes.channelize_map,
-                    &post_inf,
-                    &post_des,
-                ),
-            ];
-
-            for (pair, map, input_ids, output_ids) in pairs {
-                let edges = map.edges();
-                assert!(
-                    !edges.is_empty(),
-                    "{name}: the {pair} fold derived no edges at all",
-                );
-                for (u, d) in &edges {
-                    assert!(
-                        input_ids.contains(u),
-                        "{name}: {u:?} is an edge origin at {pair} but not an input-pane id",
-                    );
-                    assert!(
-                        output_ids.contains(&d.id),
-                        "{name}: {:?} is an edge target at {pair} but not an output-pane id",
-                        d.id,
-                    );
-                }
-                // Dense: a node present in both panes is its own self-edge, and
-                // a node descends from itself — so a consumer only ever follows
-                // edges, never reconstructs one, and reads ancestry off the
-                // label it finds there.
-                for id in input_ids.intersection(output_ids) {
-                    let self_edge = map.upstream(id).iter().find(|l| l.id == *id);
-                    assert!(
-                        self_edge.is_some_and(|l| l.labels.has_ancestry()),
-                        "{name}: {id:?} survives {pair} without an ancestry self-edge",
-                    );
-                }
-                // A non-self edge is the only proof a row was consulted: a
-                // pane pair whose phases rewrote nothing derives its whole
-                // map from the two pane id sets.
-                if edges.iter().any(|(u, d)| *u != d.id) {
-                    exercised.push(format!("{name} / {pair}"));
-                }
-            }
-        }
-        exercised.sort();
-        assert_eq!(
-            exercised, EXERCISED_BOUNDARIES,
-            "the pane pairs at which the fold actually reads a row have changed",
-        );
-    }
-
-    /// The `program / pane pair` names at which a **blame** edge reaches the
-    /// provenance map — where a rewrite named blame and the fold labelled the
-    /// edge it contributed. See
-    /// [`blame_reaches_the_provenance_map_labelled`].
-    const RELATING_BOUNDARIES: &[&str] = &[
-        "for_accumulator / post-inference → post-channelize",
-        "transaction / post-inference → post-channelize",
-    ];
-
-    /// **Blame reaches the provenance map, labelled**: the `blame` column is
-    /// closed transitively alongside `parents`, so a consumer receives the
-    /// blame edges and can render or prune them.
-    ///
-    /// Pinned as the set of pane pairs where such an edge exists, for the same
-    /// reason [`EXERCISED_BOUNDARIES`] is pinned: blame is named at a handful of
-    /// sites, all in the mutability phases and so all inside the second pane
-    /// pair, and a corpus or recording edit that stopped exercising them
-    /// would otherwise leave the labelled half of the map untested.
-    ///
-    /// A blame edge is *only* blame here: no corpus rewrite both
-    /// consumes a node and blames it, so nothing in the corpus pins the
-    /// both-labels case — the fold tests in `provenance.rs` do.
-    #[test]
-    fn blame_reaches_the_provenance_map_labelled() {
-        let mut relating: Vec<String> = Vec::new();
-        for (name, code) in corpus() {
-            let program = compile_ok(&code);
-            let panes = program.materialize_panes();
-            let pairs = [
-                ("pre-inference → post-inference", &panes.infer_map),
-                ("post-inference → post-channelize", &panes.channelize_map),
-            ];
-            for (pair, map) in pairs {
-                if map.edges().iter().any(|(_, d)| d.labels.has_blame()) {
-                    relating.push(format!("{name} / {pair}"));
-                }
-            }
-        }
-        relating.sort();
-        assert_eq!(
-            relating, RELATING_BOUNDARIES,
-            "the pane pairs at which blame contributes an edge have changed",
-        );
-    }
-
-    /// The corpus programs whose definitions inference **generalizes and then
-    /// specializes** — the ones monomorphization actually clones a subtree for.
-    /// Everything else in the corpus is first-order, and mono mints nothing.
-    const SPECIALIZING: &[&str] = &["generator_pipeline", "udf_chain"];
-
-    /// Monomorphization — the one thing that mints between the first two panes —
-    /// explains every node it produces, on first-order and specializing
-    /// programs alike.
-    ///
-    /// Two recordings get it there, and both are needed: `specialize_use` sinks
-    /// the clone's `on_copy` pairs, and `coalesce_generalized_let` sinks the
-    /// chain of `let`s the binding rebuilds itself as. Without the second, a
-    /// specializing program leaves one unrecorded `let` per demanded
-    /// specialization — a per-program count that tracks how many types the body
-    /// asked for, which is why it read as a small constant on this corpus.
-    ///
-    /// Asserted both ways so the zero cannot be vacuous: a specializing program
-    /// must also *kill* nodes here, since its generalized definition is
-    /// replaced by clones. A regression that stopped running mono at all would
-    /// otherwise pass.
-    #[test]
-    fn monomorphization_explains_every_node_it_produces() {
-        for (name, code) in corpus() {
-            let panes = compile_ok(&code).materialize_panes();
-            let c = LeakCounts::tally(&panes.infer_leaks);
-            assert_eq!(c.dangling_parent, 0, "{name}: {c:?}");
-            assert_eq!(
-                c.unrecorded, 0,
-                "{name}: pre-inference → post-inference is uncaptured: {c:?}"
-            );
-            if SPECIALIZING.contains(&name) {
-                assert!(
-                    !panes.infer_deaths.is_empty(),
-                    "{name}: specializes, so the generalized definition must die"
-                );
-            }
-        }
-    }
-
-    /// All four phases inside the second pane pair explain what they produce.
-    ///
-    /// The interesting programs are the ones that drive a *whole-program*
-    /// rewrite, where naming one node is least obviously applicable: a
-    /// transaction is disassembled into a commit carrier whose pieces have no
-    /// single source node, and a defer cluster becomes a `LetRec` assembled from
-    /// contributions scattered across the body. Both are covered by recording
-    /// against the node each product stands in for — the `with begin():`
-    /// statement, the register declaration, the `let d = Defer`.
-    ///
-    /// Asserted with a non-vacuity guard for the same reason the first pane pair is: a
-    /// program that reaches one of these phases must also kill nodes, so a
-    /// regression that stopped running the phase cannot pass as capture.
-    #[test]
-    fn the_second_pane_pair_explains_every_node_its_phases_produce() {
-        /// Reaches `transact_phase` or `channelize` — the whole-program
-        /// rewrites, and the last two phases to adopt the recorder.
-        const WHOLE_PROGRAM_REWRITES: &[&str] = &["transaction", "feed_loop", "generator_pipeline"];
-        for (name, code) in corpus() {
-            let panes = compile_ok(&code).materialize_panes();
-            let c = LeakCounts::tally(&panes.channelize_leaks);
-            assert_eq!(c.dangling_parent, 0, "{name}: {c:?}");
-            assert_eq!(
-                c.unrecorded, 0,
-                "{name}: post-inference → post-channelize is uncaptured: {c:?}"
-            );
-            if WHOLE_PROGRAM_REWRITES.contains(&name) {
-                assert!(
-                    !panes.channelize_deaths.is_empty(),
-                    "{name}: rewrites its whole shape, so nodes must die"
-                );
-            }
-        }
-    }
-
-    /// Deaths are the live-set difference and nothing else: what the fold reports
-    /// between two panes is exactly `input_ids ∖ output_ids` on a real program,
-    /// with no phase having declared any of them. `for_accumulator` folds a
-    /// mutation loop into a `LetRec`, so the difference is non-empty.
-    #[test]
-    fn deaths_between_two_panes_are_the_set_difference() {
-        let program = compile_ok(include_str!(
-            "../../tests/programs/for_accumulator/program.cambra"
-        ));
-        let panes = program.materialize_panes();
-        let input = collect_tree_ids(&program.post_inference_ir);
-        let output = collect_tree_ids(&program.post_channelize_ir);
-        let mut expected: Vec<NodeId> = input.difference(&output).copied().collect();
-        expected.sort_unstable();
-        assert!(!expected.is_empty(), "the fixture must actually kill nodes");
-        assert_eq!(panes.channelize_deaths, expected);
-    }
-
-    /// **Distinct predicate terms never share a `NodeId`** — with each other, or
-    /// with the main tree — on the three trees the panes retain.
-    ///
-    /// [`assert_unique_node_ids`] runs the same [`predicate_id_collisions`] walk
-    /// at every phase boundary, so what this adds is the panes: `pre_inference_ir`
-    /// is snapshotted after `uniquify`, between two boundaries, and no boundary
-    /// walk reaches it.
-    ///
-    /// What the walk catches is a rebuild that **preserves ids when it should
-    /// not**. A predicate cannot be mutated through its `Rc`, so every rewrite
-    /// builds a new `Rc` and repoints the refinement it was handed. That is a
-    /// *replacement* only if the walk reaches every occurrence; otherwise the
-    /// original survives on some type the walk missed, and preserving ids puts one
-    /// id-set on two live terms. `PredMemo::replacing` is the opt-in for walks
-    /// that do reach everything, and `uniquify` — the only one — asserts its own
-    /// 1:1 correspondence separately.
-    #[test]
-    fn distinct_predicate_terms_never_share_a_node_id() {
-        let mut found: Vec<(String, usize, &'static str)> = Vec::new();
-        for (name, code) in corpus() {
-            let program = compile_ok(&code);
-            for (pane, tree) in [
-                ("pre-inference", &program.pre_inference_ir),
-                ("post-inference", &program.post_inference_ir),
-                ("post-channelize", &program.post_channelize_ir),
-            ] {
-                for (_, kind) in predicate_id_collisions(tree) {
-                    found.push((format!("{name} / {pane}"), 1, kind));
-                }
-            }
-        }
-        assert!(
-            found.is_empty(),
-            "{} predicate id collisions: {found:?}",
-            found.len(),
-        );
-    }
-
-    /// A phase that rewrites the program records its rewrites under its own phase
-    /// tag — the tag being the one part of a row no recording site knows, and the
-    /// only thing that places a row in a pane pair's fold.
-    ///
-    /// A phase that rewrites *nothing* on a given program records nothing, which
-    /// is the preserve case and correct (most of the corpus preserves end to
-    /// end), so the fixture is one that drives three of the five: the
-    /// transaction, which `transact_phase` disassembles, `mut_elim` rebuilds as
-    /// a `LetRec`, and `channelize` rewrites.
-    #[test]
-    fn a_rewriting_phase_tags_its_rows_with_itself() {
-        let (_, code) = corpus()
-            .into_iter()
-            .find(|(name, _)| *name == "transaction")
-            .expect("the transaction fixture");
-        let program = compile_ok(&code);
-        let mut recorded = program.provenance_table.recorded_phases();
-        recorded.sort_by_key(|p| format!("{p:?}"));
-        assert_eq!(
-            recorded,
-            vec![
-                Phase::Channelize,
-                Phase::Infer,
-                Phase::Letrec,
-                Phase::Transact
-            ],
-            "the transaction fixture is rewritten by exactly these four phases",
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Perf sanity for pane capture
-    // -----------------------------------------------------------------------
-
-    /// One generated program shape, parameterized by size.
-    ///
-    /// The shapes are chosen to hit the phases the panes span: `comprehension`
-    /// and `arith` are `inline`-light and mostly id-preserving, `udf` drives
-    /// inlining and monomorphization, `loop_acc` drives `mut_elim`, and `feed`
-    /// drives `channelize`.
-    ///
-    /// Sizes are deliberately small. Compile time here is **superlinear in
-    /// program size** for reasons that predate provenance capture (a UDF-heavy 63-line
-    /// program compiles in tens of seconds), so a corpus large enough to be a
-    /// benchmark would be too slow to run; this is a sanity check on the
-    /// *ratio* between capture on and capture off, not a benchmark.
-    fn generated(shape: &str, n: usize) -> String {
-        match shape {
-            "arith" => {
-                let terms: Vec<String> = (1..=n).map(|i| format!("{i} * {}", i + 1)).collect();
-                format!("x = {}\nx\n", terms.join(" + "))
-            }
-            "comprehension" => {
-                // Independent comprehensions summed, rather than a chain: a
-                // chained comprehension trips an unrelated substitution bug
-                // ("discharged binder still free after substitution into
-                // predicate") that has nothing to do with provenance.
-                let parts: Vec<String> = (0..n)
-                    .map(|i| format!("sum([y + {i} for y in xs if y > 0])"))
-                    .collect();
-                format!("xs = [1, 2, 3, 4, 5]\nt = {}\nt\n", parts.join(" + "))
-            }
-            "udf" => {
-                let mut src = String::new();
-                for i in 0..n {
-                    src.push_str(&format!("def f{i}(x):\n    x + {i}\n"));
-                }
-                src.push_str("xs = [1, 2, 3, 4, 5]\n");
-                let calls: Vec<String> = (0..n).map(|i| format!("f{i}(y)")).collect();
-                src.push_str(&format!("[{} for y in xs]\n", calls.join(" + ")));
-                src
-            }
-            "loop_acc" => {
-                let mut src = String::from("xs = [1, 2, 3, 4, 5]\n");
-                for i in 0..n {
-                    src.push_str(&format!("a{i}: Mut(Int) := 0\n"));
-                }
-                src.push_str("for v in xs:\n");
-                for i in 0..n {
-                    src.push_str(&format!("    a{i} += v + {i}\n"));
-                }
-                src.push_str(&format!("a{}\n", n - 1));
-                src
-            }
-            "feed" => {
-                let mut src = String::from("out = defer()\nxs = [1, 2, 3, 4, 5]\n");
-                for i in 0..n {
-                    src.push_str(&format!("for v in xs:\n    out << v + {i}\n"));
-                }
-                src.push_str("out\n");
-                src
-            }
-            other => panic!("unknown shape {other}"),
-        }
-    }
-
-    /// The perf corpus: `(shape, size)` pairs, sized to keep the whole run in
-    /// the low seconds per repetition.
-    const PERF_CORPUS: &[(&str, usize)] = &[
-        ("arith", 40),
-        ("arith", 80),
-        ("arith", 160),
-        ("comprehension", 6),
-        ("comprehension", 12),
-        ("comprehension", 20),
-        ("udf", 5),
-        ("udf", 8),
-        ("udf", 12),
-        ("loop_acc", 4),
-        ("loop_acc", 8),
-        ("loop_acc", 14),
-        ("feed", 3),
-        ("feed", 6),
-        ("feed", 12),
-    ];
-
-    /// Rough compile-time and retained-memory sanity for pane capture. Ignored
-    /// by default — it is a measurement, not an assertion.
-    ///
-    /// Run it as two interleaved processes so the two arms see the same machine
-    /// state, and take the min over repetitions:
-    ///
-    /// ```text
-    /// for i in 1 2 3; do
-    ///   CAMBRA_PROVENANCE=1 cargo test --release --lib provenance_pane_perf -- --ignored --nocapture
-    ///   CAMBRA_PROVENANCE=0 cargo test --release --lib provenance_pane_perf -- --ignored --nocapture
-    /// done
-    /// ```
-    ///
-    /// With capture on it also materializes and folds both panes, since that is
-    /// the cost the design actually incurs.
-    #[test]
-    #[ignore = "measurement, not an assertion; see the doc comment for the driver"]
-    fn provenance_pane_perf() {
-        let capture = provenance_capture_enabled();
-        let reps: usize = std::env::var(PERF_REPS_ENV)
-            .ok()
-            .and_then(|v| v.parse().ok())
-            .unwrap_or(3);
-        let mut total_compile = std::time::Duration::ZERO;
-        let mut total_fold = std::time::Duration::ZERO;
-        for (shape, n) in PERF_CORPUS {
-            let code = generated(shape, *n);
-            let mut best_compile = std::time::Duration::MAX;
-            let mut best_fold = std::time::Duration::MAX;
-            let mut rows = 0usize;
-            let mut tags = 0usize;
-            // The three retained pane snapshots are unconditional — they are not
-            // part of what the capture switch turns off — so their size is the
-            // pane design's real memory floor, against which the logs are noise.
-            let mut panes_nodes = 0usize;
-            for _ in 0..reps {
-                let t0 = std::time::Instant::now();
-                let program = compile_ok(&code);
-                best_compile = best_compile.min(t0.elapsed());
-                rows = program.provenance_table.len();
-                tags = program.provenance_table.tag_count();
-                panes_nodes = collect_tree_ids(&program.pre_inference_ir).len()
-                    + collect_tree_ids(&program.post_inference_ir).len()
-                    + collect_tree_ids(&program.post_channelize_ir).len();
-                if capture {
-                    let t1 = std::time::Instant::now();
-                    let panes = program.materialize_panes();
-                    std::hint::black_box(&panes.post_channelize);
-                    best_fold = best_fold.min(t1.elapsed());
-                }
-            }
-            if !capture {
-                best_fold = std::time::Duration::ZERO;
-            }
-            total_compile += best_compile;
-            total_fold += best_fold;
-            eprintln!(
-                "[perf capture={capture}] {shape}/{n}: compile {:?} fold {:?} rows {rows} \
-                 tags {tags} pane_nodes {panes_nodes} lines {}",
-                best_compile,
-                best_fold,
-                code.lines().count(),
-            );
-        }
-        eprintln!("[perf capture={capture}] TOTAL compile {total_compile:?} fold {total_fold:?}");
-    }
 
     /// Driver that runs `compile_program` for an error-only test, returning
     /// the collected error list. Discards the program — these tests only
